add deps controller
https://online.sbis.ru/opendoc.html?guid=c805496c-3a16-485a-a7ab-b76205e80173
